src/load/requirements.txt
boto3
s3fs
pandas
psycopg2-binary